requirements.txt
Django==1.8.4
django-appconf==1.0.1
django-geoposition==0.2.2
django-mptt==0.7.4
ranking==0.3.1
six==1.9.0
wheel==0.24.0
